fe130f20 ENH: Wrap LandmarkBasedTransformInitializer
5138410f BUG: Use ParameterValueType for computation of landmark affine

+++ Modules/Registration/Common/test/itkLandmarkBasedTransformInitializerTest.cxx
@@ -674,7 +674,7 @@ int itkLandmarkBasedTransformInitializerTest(int, char * [])
   }
 
   {
-  typedef itk::Transform< double, 3, 3 > TransformType;
+  typedef itk::Transform< float, 3, 3 > TransformType;
 
   typedef itk::LandmarkBasedTransformInitializer< TransformType > TransformInitializerType;
